refactor; change issues __init__ to use categories
