+++ cognite/neat/rules/issues/importing.py
@@ -3,6 +3,19 @@ from dataclasses import dataclass
 
 from .base import NeatValidationError, ValidationWarning
 
+__all__ = [
+    "ModelImportWarning",
+    "UnknownComponentWarning",
+    "UnknownSubComponentWarning",
+    "IgnoredComponentWarning",
+    "UnknownPropertyWarning",
+    "ModelImportError",
+    "InvalidComponentError",
+    "MissingParentDefinitionError",
+    "MissingIdentifierError",
+    "UnsupportedPropertyTypeError",
+]
+
 
 @dataclass(frozen=True)
 class ModelImportWarning(ValidationWarning, ABC):
