+++ tests/tests_unit/rules/importer/test_excel2rules.py
@@ -7,7 +7,8 @@ from pydantic import ValidationError
 
 from cognite.neat.rules import importer
 from cognite.neat.rules.importers._spreadsheet2rules import ExcelImporter
-from cognite.neat.rules.issues import IssueList, SheetMissingError
+from cognite.neat.rules.issues import IssueList
+from cognite.neat.rules.issues.spreadsheet_file import SheetMissingError
 from cognite.neat.rules.models import Tables
 from cognite.neat.rules.models._rules import DomainRules, InformationRules
 from cognite.neat.rules.models._rules.base import RoleTypes
